requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.0.0
Django==4.2.15
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.15.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.1
threadpoolctl==3.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
zope.interface==7.2
psycopg2-binary
whitenoise
